requirements.txt
pytest==7.4.3
selene==2.0.0rc8
selenium~=4.11.2
pydantic==2.5.3
pydantic-settings
python-dotenv==1.0.0
allure-pytest==2.13.2